updates msg, srv, and enumeration types in theo_comm and theo_msgs to reduce message sizes, provides include guards where missing across provided packages

+++ src/theo_autoop/include/controller_modes.hpp
@@ -1,7 +1,9 @@
 #pragma once
 
-enum class ControllerMode : int {
+#include <cstdint>
+
+enum class ControllerMode : int8_t {
     Idle,       // doing nothing
-    Responding, // responding to some requested pose, generates feedback when response completes
+    Responding, // responding to some requested pose, generates feedback to brokerage when response completes
     Active      // processing commands on topic
 };
